# packages/contracts/openapi/wedge_openapi.yaml
openapi: 3.0.3
info:
  title: Wedge API
  version: 1.0.0-draft
  description: Wedge public REST API (/api), internal callback API (/internal), and canonical transport schemas. This draft
    includes report shares, checkpoint callback, and evidence packet retrieval.
servers:
- url: http://localhost:8080
  description: local API server for Swagger MVP testing
- url: https://api.wedge.example.com
  description: production placeholder
- url: https://staging-api.wedge.example.com
  description: staging placeholder
security:
- HumanBearer: []
...
paths:
...
  /api/runs/{runId}/artifacts:
    get:
      tags:
      - Artifacts
      operationId: listRunArtifacts
      summary: 실행 아티팩트 목록 조회
      description: 스크린샷/HTML/trace 같은 저장 아티팩트 목록과 content URL을 확인한다. Runner callback 이후 결과 확인에 사용한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: type
        in: query
        schema:
          type: string
      - name: stepId
        in: query
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Run artifacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListArtifactResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /internal/runner/runs/{runId}/accepted:
    post:
      tags:
      - InternalRunner
      operationId: runnerAccepted
      summary: Runner 실행 수락 callback
      description: Runner가 실행 요청을 수락했음을 API 서버에 알리는 내부 callback이다. INTERNAL 토큰과 X-Event-Id/X-Worker-Id 헤더로 중복 처리까지 테스트한다.
      security:
      - InternalBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: X-Worker-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Event-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Signature
        in: header
        description: HMAC-SHA256 digest of the raw JSON request body using the shared runner callback signature secret. The preferred format is `hmac-sha256=<hex>`; raw hex is accepted for compatibility.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunnerAcceptedRequest'
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /internal/runner/runs/{runId}/step-events:
    post:
      tags:
      - InternalRunner
      operationId: runnerStepEvents
      summary: Runner 단계 이벤트 callback
      description: Runner가 step 시작/완료 같은 이벤트를 밀어 넣는 내부 callback이다. Run timeline/evidence 저장 흐름 검증에 사용한다.
      security:
      - InternalBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: X-Worker-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Event-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Signature
        in: header
        description: HMAC-SHA256 digest of the raw JSON request body using the shared runner callback signature secret. The preferred format is `hmac-sha256=<hex>`; raw hex is accepted for compatibility.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunnerStepEventsRequest'
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /internal/runner/runs/{runId}/control-state:
    get:
      tags:
      - InternalRunner
      operationId: runnerControlState
      summary: Runner 실행 제어 상태 조회
      description: Runner가 step 사이에서 STOP_REQUESTED 상태를 소비할 수 있도록 현재 Run 제어 상태를 조회한다.
      security:
      - InternalBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: X-Worker-Id
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Current control state
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/RunnerControlStateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /internal/runner/runs/{runId}/artifacts:
    post:
      tags:
      - InternalRunner
      operationId: runnerArtifacts
      summary: Runner 아티팩트 callback
      description: Runner가 캡처한 파일/스크린샷 메타데이터를 저장하는 내부 callback이다. 이후 artifacts API에서 조회되는지 테스트한다.
      security:
      - InternalBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: X-Worker-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Event-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Signature
        in: header
        description: HMAC-SHA256 digest of the raw JSON request body using the shared runner callback signature secret. The preferred format is `hmac-sha256=<hex>`; raw hex is accepted for compatibility.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunnerArtifactsRequest'
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /internal/runner/runs/{runId}/finished:
    post:
      tags:
      - InternalRunner
      operationId: runnerFinished
      summary: Runner 완료 callback
      description: Runner가 Run 완료를 알리고 상태를 COMPLETED로 전환하는 내부 callback이다. 분석 요청 전 완료 상태 만들기에 사용한다.
      security:
      - InternalBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: X-Worker-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Event-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Signature
        in: header
        description: HMAC-SHA256 digest of the raw JSON request body using the shared runner callback signature secret. The preferred format is `hmac-sha256=<hex>`; raw hex is accepted for compatibility.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunnerFinishedRequest'
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /internal/runner/runs/{runId}/failed:
    post:
      tags:
      - InternalRunner
      operationId: runnerFailed
      summary: Runner 실패 callback
      description: Runner 실패 사유를 저장하고 Run 상태를 FAILED로 전환하는 내부 callback이다. 실패 처리/에러 메시지 검증에 사용한다.
      security:
      - InternalBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: X-Worker-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Event-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Signature
        in: header
        description: HMAC-SHA256 digest of the raw JSON request body using the shared runner callback signature secret. The preferred format is `hmac-sha256=<hex>`; raw hex is accepted for compatibility.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunnerFailedRequest'
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /internal/runner/runs/{runId}/agent-events:
    post:
      tags:
      - InternalRunner
      operationId: runnerAgentEvents
      summary: Runner Agent 이벤트 callback
      description: Runner Agent Runtime이 관찰/결정/정책/검증 이벤트 batch를 API 서버에 알리는 내부 callback이다.
      security:
      - InternalBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: X-Worker-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Event-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Signature
        in: header
        description: HMAC-SHA256 digest of the raw JSON request body using the shared runner callback signature secret. The preferred format is `hmac-sha256=<hex>`; raw hex is accepted for compatibility.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunnerAgentEventsRequest'
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /internal/runner/runs/{runId}/agent-traces:
    post:
      tags:
      - InternalRunner
      operationId: runnerAgentTrace
      summary: Runner Agent trace callback
      description: Runner Agent Runtime이 최종 AgentTrace를 API 서버에 알리는 내부 callback이다. TRACE artifact와 같은 trace 계약을 사용한다.
      security:
      - InternalBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: X-Worker-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Event-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Signature
        in: header
        description: HMAC-SHA256 digest of the raw JSON request body using the shared runner callback signature secret. The preferred format is `hmac-sha256=<hex>`; raw hex is accepted for compatibility.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunnerAgentTraceRequest'
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /internal/runner/runs/{runId}/checkpoints:
    post:
      tags:
      - InternalRunner
      operationId: runnerCheckpoints
      summary: Runner 체크포인트 callback
      description: Runner가 checkpoint와 observation batch를 저장하는 핵심 내부 callback이다. evidence-packet과 report 입력 데이터가 생성되는지 확인한다.
      security:
      - InternalBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      - name: X-Worker-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Event-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Signature
        in: header
        description: HMAC-SHA256 digest of the raw JSON request body using the shared runner callback signature secret. The preferred format is `hmac-sha256=<hex>`; raw hex is accepted for compatibility.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunnerCheckpointsRequest'
      responses:
        '202':
          description: Checkpoint batch accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.wedge.example.com/oauth2/authorize
          tokenUrl: https://auth.wedge.example.com/oauth2/token
          scopes:
            wedge.read: Read Wedge resources
            wedge.execute: Create and execute runs
            wedge.export: Export reports
            wedge.admin: Administrative actions
        clientCredentials:
          tokenUrl: https://auth.wedge.example.com/oauth2/token
          scopes:
            wedge.read: Read Wedge resources
            wedge.execute: Create and execute runs
            wedge.export: Export reports
            wedge.admin: Administrative actions
    InternalBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    HumanBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    Unauthorized:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    AckResponse:
      type: object
      properties:
        data:
          type: object
          nullable: true
          additionalProperties: true
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
...
    RunnerAcceptedRequest:
      type: object
      properties:
        workerId:
          type: string
        acceptedAt:
          type: string
          format: date-time
        browserSessionId:
          type: string
      required:
      - workerId
      - acceptedAt
      - browserSessionId
...
    RunnerStepEventsRequest:
      type: object
      properties:
        events:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/RunnerStepEvent'
      required:
      - events
...
    RunnerArtifactsRequest:
      type: object
      properties:
        artifacts:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              artifactId:
                type: string
                format: uuid
              stepKey:
                type: string
                description: Scenario step key for step-scoped artifacts. Reserved agent run-scoped keys such as agent_trace and agent_scenario_plan_export are persisted without resolving a ScenarioPlan step.
              artifactType:
                type: string
                enum:
                - FRAME
                - SCREENSHOT
                - DOM_SNAPSHOT
                - AX_TREE
                - TRACE
                - HAR
                - CONSOLE_LOG
                - REPORT_PDF
                - REPORT_MARKDOWN
                - REPORT_HTML
                - REPORT_JSON
                - OTHER
              bucket:
                type: string
              key:
                type: string
              mimeType:
                type: string
              width:
                type: integer
                minimum: 1
                nullable: true
              height:
                type: integer
                minimum: 1
                nullable: true
              sizeBytes:
                type: integer
                minimum: 0
              sha256:
                type: string
              createdAt:
                type: string
                format: date-time
            required:
            - artifactId
            - stepKey
            - artifactType
            - bucket
            - key
            - mimeType
            - sizeBytes
            - sha256
            - createdAt
      required:
      - artifacts
    AgentEvent:
      $ref: ../schemas/agent-event.schema.json
...
    RunnerAgentEventsRequest:
      type: object
      properties:
        events:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AgentEvent'
      required:
      - events
...
    RunnerFinishedRequest:
      type: object
      properties:
        workerId:
          type: string
        executionFinishedAt:
          type: string
          format: date-time
        summary:
          type: object
          properties:
            completedStepCount:
              type: integer
              minimum: 0
            failedStepCount:
              type: integer
              minimum: 0
            stopped:
              type: boolean
              description: True when runner execution stopped at a planned safety stop condition or after an explicit stop request. Planned scenario stops still complete the run; only a prior STOP_REQUESTED run becomes STOPPED.
          required:
          - completedStepCount
          - failedStepCount
          - stopped
      required:
      - workerId
      - executionFinishedAt
      - summary
    RunnerFailedRequest:
      type: object
      properties:
        workerId:
          type: string
        failedAt:
          type: string
          format: date-time
        failureCode:
          type: string
        failureMessage:
          type: string
        resultCompleteness:
          type: string
          enum:
          - NONE
          - PARTIAL
          - FINAL
        summary:
          type: object
          properties:
            completedStepCount:
              type: integer
              minimum: 0
            failedStepCount:
              type: integer
              minimum: 0
            stopped:
              type: boolean
          required:
          - completedStepCount
          - failedStepCount
          - stopped
        failureArtifactRefs:
          type: array
          items:
            type: string
      required:
      - workerId
      - failedAt
      - failureCode
      - failureMessage
      - resultCompleteness
    RunnerControlStateResponse:
      type: object
      properties:
        runId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - CREATED
          - QUEUED
          - STARTING
          - RUNNING
          - STOP_REQUESTED
          - STOPPED
          - COMPLETED
          - FAILED
        stopRequested:
          type: boolean
        resultCompleteness:
          type: string
          enum:
          - NONE
          - PARTIAL
          - FINAL
          nullable: true
      required:
      - runId
      - status
      - stopRequested
...
    ListArtifactResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Artifact'
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
...